+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* v0.1.1
+	* Better acceptance tests
+	* More validation
+	* Use of genfile and/or phasefile
 * v0.1.0
 	* First versionned beta release
 * v0.0.0